export/requirements.txt
docxtpl==0.14.2
icecream==2.1.1
python-docx==0.8.11
python-frontmatter==1.0.0
PyYAML==6.0